# openapi.yml
openapi: 3.1.0
info:
  title: Neue Konzepte BaaS Backend API
  version: 0.1.0
  description: |
    Backend API for the Neue Konzepte BaaS project.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/crops/{cropID}/season:
    put:
      tags: [crops, seasons]
      summary: Tie a crop to a season
      description: |
        Assigns the crop's default season rule (admin) or the caller's own
        farm's rule (farmer), creating it if none exists yet or repointing it
        at the given season otherwise. Renting the crop is then only allowed
        inside that season's window — see `POST /api/rentals`.

        An admin's season must be one of the defaults; a farmer's must belong
        to their own farm.
      operationId: assignCropSeason
      parameters:
        - $ref: '#/components/parameters/CropID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCropSeasonRequest'
      responses:
        '200':
          description: The crop's season rule, created or repointed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CropSeasonRule'
        '400':
          description: Malformed body, invalid crop id, or invalid season id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidBody:
                  value:
                    error: invalid request body
                invalidCropID:
                  value:
                    error: invalid crop id
                invalidSeasonID:
                  value:
                    error: invalid season id
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: The authenticated account is not an admin or farmer, a farmer has no farm, or the season does not belong to the caller's set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: season does not belong to the caller's set
        '404':
          description: No crop or season with the given id exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: crop or season not found
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [crops, seasons]
      summary: Remove a crop's season rule
      description: |
        Drops the crop's default rule (admin) or the caller's own farm's rule
        (farmer), reverting it to the default rule if one exists, or to
        unrestricted otherwise.
      operationId: removeCropSeasonRule
      parameters:
        - $ref: '#/components/parameters/CropID'
      responses:
        '204':
          description: Rule removed.
        '400':
          description: Invalid crop id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: invalid crop id
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: The authenticated account is not an admin or farmer, or a farmer has no farm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: insufficient permissions
        '404':
          description: The crop has no season rule in the caller's set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: crop has no season rule in the caller's set
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/seasons:
    get:
      tags: [seasons]
      summary: List the caller's seasons
      description: |
        Returns the default set (admin) or the union of the default set and
        the caller's own farm's seasons (farmer) — everything they could pick
        when assigning a crop's season. Requires the `admin` or `farmer` role.
      operationId: getSeasons
      responses:
        '200':
          description: The caller's seasons.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Season'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: The authenticated account is not an admin or farmer, or a farmer has no farm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: insufficient permissions
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [seasons]
      summary: Add a season
      description: |
        Adds a season to the default set (admin) or the caller's own farm's
        seasons (farmer) — a farmer may create and manage their own seasons at
        any time, independent of the defaults and of any crop. Requires the
        `admin` or `farmer` role.
      operationId: createSeason
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeasonRequest'
      responses:
        '201':
          description: Season created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Season'
        '400':
          description: Malformed body, missing name, or a month/day outside its valid range.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidBody:
                  value:
                    error: invalid request body
                missingName:
                  value:
                    error: name is required
                invalidMonth:
                  value:
                    error: startMonth and endMonth must be between 1 and 12
                invalidDay:
                  value:
                    error: startDay and endDay must be between 1 and 31
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: The authenticated account is not an admin or farmer, or a farmer has no farm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: insufficient permissions
        '500':
          $ref: '#/components/responses/InternalError'
  /api/seasons/{seasonID}:
    put:
      tags: [seasons]
      summary: Update a season
      description: |
        Rewrites an existing season. An admin may edit any default season, a
        farmer only their own farm's. Requires the `admin` or `farmer` role.
      operationId: updateSeason
      parameters:
        - name: seasonID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSeasonRequest'
      responses:
        '200':
          description: Season updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Season'
        '400':
          description: Malformed body, invalid season id, missing name, or a month/day outside its valid range.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidBody:
                  value:
                    error: invalid request body
                invalidSeasonID:
                  value:
                    error: invalid season id
                missingName:
                  value:
                    error: name is required
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: The authenticated account is not an admin or farmer, or a farmer has no farm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: insufficient permissions
        '404':
          description: No season with this id is visible to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: season not found
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [seasons]
      summary: Delete a season
      description: |
        Removes one season, under the same ownership rules as
        `PUT /api/seasons/{seasonID}`. Requires the `admin` or `farmer` role.
      operationId: deleteSeason
      parameters:
        - name: seasonID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Season deleted.
        '400':
          description: Invalid season id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: invalid season id
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: The authenticated account is not an admin or farmer, or a farmer has no farm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: insufficient permissions
        '404':
          description: No season with this id is visible to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: season not found
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
...
    CropID:
      name: cropID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthenticated:
      description: No valid session cookie.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not authenticated
...
    InternalError:
      description: Internal error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: internal error
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly JWT cookie set by `POST /api/auth/login`. Valid for 15 minutes.
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    Season:
      type: object
      description: |
        A recurring yearly calendar window - "Spring: Mar 1 - May 31" - that
        repeats every year regardless of which specific year a rental falls
        in. There is one default set, which an admin maintains and every farm
        starts from; a farmer may also create, edit and delete their own
        seasons at any time, independent of the defaults and of any crop. A
        crop's season rule (see `CropSeasonRule`) points at one of these rows.
      required: [id, name, farmId, startMonth, startDay, endMonth, endDay, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Spring
        farmId:
          type: string
          format: uuid
          nullable: true
          description: Null for one of the global defaults, the farm's id for that farm's own version.
        startMonth:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        startDay:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
        endMonth:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        endDay:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
          description: |
            A season whose endMonth/endDay is earlier in the year than its
            startMonth/startDay wraps the new year (e.g. Winter: Dec 1 - Feb 28).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSeasonRequest:
      type: object
      required: [name, startMonth, startDay, endMonth, endDay]
      properties:
        name:
          type: string
          description: Surrounding whitespace is trimmed.
          example: Spring
        startMonth:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        startDay:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
        endMonth:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        endDay:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
    UpdateSeasonRequest:
      type: object
      required: [name, startMonth, startDay, endMonth, endDay]
      properties:
        name:
          type: string
          example: Spring
        startMonth:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        startDay:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
        endMonth:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        endDay:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
    CropSeasonRule:
      type: object
      description: |
        Ties a crop to a season, so renting it is only allowed inside that
        season's window. The default rule (farmId null) applies to every farm
        that has not overridden it; a farm's own rule (farmId set) points at
        one of that farm's own seasons instead. A crop with no rule at all,
        for a farm or as a default, is unrestricted / year-round.
      required: [id, cropId, seasonId, farmId]
      properties:
        id:
          type: string
          format: uuid
        cropId:
          type: string
          format: uuid
        seasonId:
          type: string
          format: uuid
        farmId:
          type: string
          format: uuid
          nullable: true
          description: Null for the default rule, the farm's id for that farm's own rule.
    AssignCropSeasonRequest:
      type: object
      required: [seasonId]
      properties:
        seasonId:
          type: string
          format: uuid
          description: |
            Must be one of the defaults (admin) or belong to the caller's own
            farm (farmer).
...
security:
  - accessTokenCookie: []